requirements.txt
gradio
faiss-cpu
sentence-transformers
transformers
accelerate
bitsandbytes
torch